Add test runner to run the test in the vm.
Run the test in the vm and collect the test result. Refer to the logic of
ssh_run.rb

+++ src/main/java/org/kafkaliu/test/vagrant/VagrantRunAfters.java
@@ -3,11 +3,13 @@ package org.kafkaliu.test.vagrant;
 import java.util.ArrayList;
 import java.util.List;
 
+import org.junit.runner.RunWith;
 import org.junit.runners.model.MultipleFailureException;
 import org.junit.runners.model.Statement;
-import org.kafkaliu.test.vagrant.annotations.VagrantConfigure;
 import org.kafkaliu.test.vagrant.ruby.VagrantCli;
 import org.kafkaliu.test.vagrant.ruby.VagrantEnvironment;
+import org.kafkaliu.test.vagrant.server.VagrantServerTestRunner;
+import org.kafkaliu.test.vagrant.server.annotations.VagrantConfigure;
 
 public class VagrantRunAfters extends Statement {
 	
@@ -36,8 +38,10 @@ public class VagrantRunAfters extends Statement {
         	if (annotation != null && annotation.needDestroyVmAfterClassTest()) {
         		cli.destroy();
         	} else {
-	        	cli.ssh("killall java");
-	        	Thread.sleep(5 * 1000);
+        		if (klass.getAnnotation(RunWith.class).annotationType().isAssignableFrom(VagrantServerTestRunner.class)) {
+		        	cli.ssh("killall java");
+		        	Thread.sleep(5 * 1000);
+        		}
         	}
         }
         MultipleFailureException.assertEmpty(errors);
